// src/interface/product.interface.ts
interface ProductReview {
    rating: number;
    comment: string;
    date: string;
    reviewerName: string;
    reviewerEmail: string;
}

interface ProductDimensions {
    width: number;
    height: number;
    depth: number;
}

interface ProductMeta {
    createdAt: string;
    updatedAt: string;
    barcode: string;
    qrCode: string;
}

export interface IProduct {
    id: number;
    title: string;
    description: string;
    category: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    tags: string[];
    brand: string;
    sku: string;
    weight: number;
    dimensions: ProductDimensions;
    warrantyInformation: string;
    shippingInformation: string;
    availabilityStatus: string;
    reviews: ProductReview[];
    returnPolicy: string;
    minimumOrderQuantity: number;
    meta: ProductMeta;
    images: string[];
    thumbnail: string;
}


export interface IProductResponse {
    skip: number;
    limit: number;
    total: number;
    products: IProduct[]
}

export interface IProductInitialState {
    products: IProduct[];
    loading: boolean;
    error: string | null;
    selectedProduct: IProduct | null;
}